+++ Classes/Tools/Statistics.cpp
@@ -39,7 +39,7 @@ void Statistics::sendStatisticsData(int type){
 		code.append("lmyspread");
 
 		MD5 md;
-		code = md.GetMd5(code.c_str(), code.length());
+		code = md.GetMd5(code.c_str(), (int)code.length());
 
 		postData.append(code.c_str());
 		postData.append("&machineCode=");
@@ -51,7 +51,7 @@ void Statistics::sendStatisticsData(int type){
 		url += "AccountsRegister";
 		//////////////////////////////////////////////////////////////////////////
 		postData = "gameID=";
-		postData.append(CCString::createWithFormat("%ld", DataModel::sharedDataModel()->userInfo->dwGameID)->getCString());
+		postData.append(CCString::createWithFormat("%d", DataModel::sharedDataModel()->userInfo->dwGameID)->getCString());
 		postData.append("&sessionID=");
 		postData.append(k_session_id);
 		postData.append("&code=");
@@ -61,7 +61,7 @@ void Statistics::sendStatisticsData(int type){
 		code.append("lmyspread");
 
 		MD5 md;
-		code = md.GetMd5(code.c_str(), code.length());
+		code = md.GetMd5(code.c_str(), (int)code.length());
 
 		postData.append(code.c_str());
 		postData.append("&machineCode=");
@@ -73,7 +73,7 @@ void Statistics::sendStatisticsData(int type){
 		url += "AccountsLogon";
 		//////////////////////////////////////////////////////////////////////////
 		postData = "gameID=";
-		postData.append(CCString::createWithFormat("%ld", DataModel::sharedDataModel()->userInfo->dwGameID)->getCString());
+		postData.append(CCString::createWithFormat("%d", DataModel::sharedDataModel()->userInfo->dwGameID)->getCString());
 		postData.append("&sessionID=");
 		postData.append(k_session_id);
 		postData.append("&code=");
@@ -83,7 +83,7 @@ void Statistics::sendStatisticsData(int type){
 		code.append("lmyspread");
 
 		MD5 md;
-		code = md.GetMd5(code.c_str(), code.length());
+		code = md.GetMd5(code.c_str(), (int)code.length());
 
 		postData.append(code.c_str());
 		postData.append("&machineCode=");
@@ -97,7 +97,7 @@ void Statistics::sendStatisticsData(int type){
 	
 	
 	request->setUrl(url.c_str());
-	request->setRequestData(postData.c_str(), strlen(postData.c_str()));
+	request->setRequestData(postData.c_str(), (int)strlen(postData.c_str()));
 	CCHttpClient::getInstance()->send(request);
 	request->release();
 }
